+++ examples/TextInput/index.js
@@ -6,9 +6,11 @@ import TextInput, { PureTextInput } from 'src/TextInput';
 
 import BasicUsage from './BasicUsage';
 import WithStatus from './WithStatus';
+import MultiLines from './MultiLines';
 
 storiesOf('TextInput', module)
     .addWithInfo('Basic usage', BasicUsage)
     .addWithInfo('With status', WithStatus)
+    .addWithInfo('Multiple lines', MultiLines)
     // Props table
     .addPropsTable(() => <TextInput />, [PureTextInput]);