+++ CHANGELOG.md
@@ -7,10 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add Apache License 2.0 for open-source.
-
-### Changed
 - Add `Installation` and `Usage` contents in README.
-- A new `<EditableBasicRow>` containing input logics is split from `<EditableText>`. (#63)
+- A new `<EditableBasicRow>` containing input logics is split from `<EditableText>`. (#63) Also supports choosing from `input` or `textarea` for its inner tag. (#64)
 
 ### Changed
 - `<EditableText>` is simplified to only hold status-related logic. (#63)